Refactor <TestDetails/> test case
